requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
azure-keyvault-secrets==4.9.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
pandas==2.2.3